[refactor] Fixed-width integral types for euclid fn

+++ lib/euclidean/euclidean.c
@@ -57,8 +57,8 @@ uint16_t euclid(uint8_t length, uint8_t density, uint8_t offset) {
   int pauses = length - density;
   int per_pulse = pauses / density;
   int remainder = pauses % density;
-  unsigned int workbeat[length];
-  unsigned int outbeat;
+  uint16_t workbeat[length];
+  uint16_t outbeat;
   uint16_t outbeat2;
   int workbeat_count = length;
